Split MS-CSSP and MS-RDPBCGR protocols from rdp_check.py
