+++ impacket/itu/rdpbcgr.py
@@ -1,0 +1,105 @@
+#!/usr/bin/env python
+# Impacket - Collection of Python classes for working with network protocols.
+#
+# Copyright (C) 2023 Fortra. All rights reserved.
+#
+# This software is provided under a slightly modified version
+# of the Apache Software License. See the accompanying LICENSE file
+# for more information.
+#
+# Description:
+#   [MS-RDPBCGR] Remote Desktop Protocol: Basic Connectivity and Graphics Remoting partial implementation
+#
+# Author:
+#  Alberto Solino (@agsolino)
+#  Alex Romero (@NtAlexio2)
+#
+
+from impacket.structure import Structure
+
+TDPU_CONNECTION_REQUEST  = 0xe0
+TPDU_CONNECTION_CONFIRM  = 0xd0
+TDPU_DATA                = 0xf0
+TPDU_REJECT              = 0x50
+TPDU_DATA_ACK            = 0x60
+
+# RDP_NEG_REQ constants
+TYPE_RDP_NEG_REQ = 1
+PROTOCOL_RDP     = 0
+PROTOCOL_SSL     = 1
+PROTOCOL_HYBRID  = 2
+
+# RDP_NEG_RSP constants
+TYPE_RDP_NEG_RSP = 2
+EXTENDED_CLIENT_DATA_SUPPORTED = 1
+DYNVC_GFX_PROTOCOL_SUPPORTED   = 2
+
+# RDP_NEG_FAILURE constants
+TYPE_RDP_NEG_FAILURE                  = 3
+SSL_REQUIRED_BY_SERVER                = 1
+SSL_NOT_ALLOWED_BY_SERVER             = 2
+SSL_CERT_NOT_ON_SERVER                = 3
+INCONSISTENT_FLAGS                    = 4
+HYBRID_REQUIRED_BY_SERVER             = 5
+SSL_WITH_USER_AUTH_REQUIRED_BY_SERVER = 6
+
+class TPKT(Structure):
+    commonHdr = (
+        ('Version','B=3'),
+        ('Reserved','B=0'),
+        ('Length','>H=len(TPDU)+4'),
+        ('_TPDU','_-TPDU','self["Length"]-4'),
+        ('TPDU',':=""'),
+    )
+
+class TPDU(Structure):
+    commonHdr = (
+        ('LengthIndicator','B=len(VariablePart)+1'),
+        ('Code','B=0'),
+        ('VariablePart',':=""'),
+    )
+
+    def __init__(self, data = None):
+        Structure.__init__(self,data)
+        self['VariablePart']=''
+
+class CR_TPDU(Structure):
+    commonHdr = (
+        ('DST-REF','<H=0'),
+        ('SRC-REF','<H=0'),
+        ('CLASS-OPTION','B=0'),
+        ('Type','B=0'),
+        ('Flags','B=0'),
+        ('Length','<H=8'),
+    )
+
+class DATA_TPDU(Structure):
+    commonHdr = (
+        ('EOT','B=0x80'),
+        ('UserData',':=""'),
+    )
+
+    def __init__(self, data = None):
+        Structure.__init__(self,data)
+        self['UserData'] =''
+
+
+class RDP_NEG_REQ(CR_TPDU):
+    structure = (
+        ('requestedProtocols','<L'),
+    )
+    def __init__(self,data=None):
+        CR_TPDU.__init__(self,data)
+        if data is None:
+            self['Type'] = TYPE_RDP_NEG_REQ
+
+class RDP_NEG_RSP(CR_TPDU):
+    structure = (
+        ('selectedProtocols','<L'),
+    )
+
+class RDP_NEG_FAILURE(CR_TPDU):
+    structure = (
+        ('failureCode','<L'),
+    )
+
